requirements.txt
openai
python-dotenv
streamlit
langchain
langchain_openai
langchain_core
langchain_community
chardet
chromadb
streamlit-chat
audio-recorder-streamlit
streamlit-float